Refactored hp into combat
